+++ verik-compiler/src/main/kotlin/io/verik/compiler/core/common/Core.kt
@@ -86,6 +86,7 @@ object Core {
 
             val F_unaryPlus = CoreKtInt.F_unaryPlus
             val F_unaryMinus = CoreKtInt.F_unaryMinus
+            val F_rangeTo_Int = CoreKtInt.F_rangeTo_Int
             val F_plus_Int = CoreKtInt.F_plus_Int
             val F_minus_Int = CoreKtInt.F_minus_Int
             val F_times_Int = CoreKtInt.F_times_Int