+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Support `CoverGroup`, `CoverPoint`, and `CoverCross`.
 - Support cover bins for `CoverPoint` and `CoverCross`.
 - Support `fill0`, `fill1`, `fillx`, and `fillz`.
+- Support `rangeTo` operator.
 
 ## [0.1.14]
 ### Importer